Restructure the shared eslint setup as a plugin under tools with config as the consumer surface
