+++ utils/config.js
@@ -61,16 +61,6 @@ const getScriptArgs = ( script, args ) => {
 						'--config',
 						require.resolve( '../config/webpack.config.js' ),
 				  ];
-		case 'lint-js':
-			return hasProjectFile( '.eslintrc.js' ) &&
-				! hasProjectFile( 'eslint.config.js' ) &&
-				! hasArgInCLI( '--config' )
-				? [
-						...args,
-						'--config',
-						require.resolve( '../config/eslint.config.js' ),
-				  ]
-				: args;
 		default:
 			return args;
 	}
